dealing with some pre-existing broken copies

+++ apps/kavita/start.sh
@@ -1,7 +1,11 @@
 #!/usr/bin/env bash
 
 if [[ ! -f "/config/appsettings.json" ]]; then
-    cp /defaults/appsettings-init.json /config/
+    if [[ -f /defaults/appsettings-init.json ]]; then
+        cp /defaults/appsettings-init.json /config/
+    elif [[ -f /app/config/appsettings-init.json ]]; then
+        cp /app/config/appsettings-init.json /config/
+    fi
 fi
 
 cd /app
